# lib/api-zod/src/generated/types/backtestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Trade } from "./trade";

export interface BacktestResult {
  symbol: string;
  days: number;
  totalTrades: number;
  wins: number;
  losses: number;
  winRate: number;
  totalProfit: number;
  totalProfitPercent: number;
  maxDrawdown: number;
  trades: Trade[];
}