# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Signal } from "./signal";

export interface DashboardSummary {
  balance: number;
  todayPnl: number;
  todayPnlPercent: number;
  openPositions: number;
  todayTrades: number;
  winRate: number;
  lastSignal?: Signal | null;
  killSwitchActive: boolean;
  autoModeActive: boolean;
}