# lib/api-zod/src/generated/types/portfolio.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Portfolio {
  balance: number;
  initialBalance: number;
  totalPnl: number;
  totalPnlPercent: number;
  openPositions: number;
  totalTrades: number;
  winRate: number;
  todayTrades: number;
  todayPnl: number;
}