# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Trading Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  allocation: number;
  stopLossPercent: number;
  takeProfitPercent: number;
  maxTradesPerDay: number;
  minConfidence: number;
  autoMode: boolean;
  liveTrading: boolean;
  killSwitch: boolean;
  binanceApiKey?: string | null;
  binanceApiSecret?: string | null;
}